+++ src/main/java/org/pkaboo/jpa/nestedset/NodeInfo.java
@@ -14,13 +14,13 @@ package org.pkaboo.jpa.nestedset;
  * of a node in a nested set tree.
  */
 public interface NodeInfo {
-    int getId();
-    int getLeftValue();
-    int getRightValue();
-    int getLevel();
-    int getRootValue();
-    void setLeftValue(int value);
-    void setRightValue(int value);
-    void setLevel(int level);
-    void setRootValue(int value);
+    Long getId();
+    Integer getLeftValue();
+    Integer getRightValue();
+    Integer getLevel();
+    Long getRootValue();
+    void setLeftValue(Integer value);
+    void setRightValue(Integer value);
+    void setLevel(Integer level);
+    void setRootValue(Long value);
 }
